bugfix/IGD-38 reorganised folders
