+++ src/masterblaster/core/assets.lua
